#2: Added remaining features. Also added tests for the mssage and queue mapper.
git-svn-id: https://svn.inf.ethz.ch/svn/systems/asl13/trunk/user33@1117 71ed90d7-4523-4855-813e-b9b7bd3c0a65

+++ milestone01/middleware-shared/src/ch/ethz/asl/util/MessageUtils.java
@@ -178,9 +178,9 @@ public class MessageUtils {
 					String content = tokens[7];
 					
 					String[] queues = queueList.split(" ");
-					List<Integer> qList = new ArrayList<Integer>();
+					List<Long> qList = new ArrayList<Long>();
 					for(String q : queues) {
-						qList.add(Integer.parseInt(q));
+						qList.add(Long.parseLong(q));
 					}
 					
 					Message m = new Message();
